Get realpath in get_port_details
This is a slightly different approch to fixing the issue with the
way the port details are resovled on linux. Currently the symlink
is converted to a real path in the serialport.c however if I
remove this code and move it to the get_port_details in the
linux.c everything seems to work fine. I don't believe this
will effect Mac and there was a Windows guard on the code
any way so the only implementation this might cause problems
for is BSD. Otherwise this is less envasive because we aren't
updating the port struct.